Physical Object Factory
